refactor(core): share HTTP URL normalization
Reuse the HTTP(S) guard for manifest links and chat connect offers.

Correct schema comments and cover rejected connect URLs.
